\name{butterfly}
\alias{butterfly}
\docType{data}
\title{
butterfly dataset}
\description{
Synthetic dataset.
}
\usage{data(butterfly)}
\format{
A matrix with 15 rows and 2 columns.
}
\details{
The butterfly dataset motivates the need for the fuzzy approach to clustering. Two clusters are clearly visible, but unit n.8 is between the two clusters, and hence its assignment is complex for standard clustering methods.
}
\references{
Giordani, P., Ferraro, M.B., Martella, F.: An Introduction to Clustering with R. Springer, Singapore (2020) \cr
Ruspini, E.H.: Numerical methods for fuzzy clustering. Inf. Sci. 2, 319-350 (1970) 
}
\author{
Paolo Giordani, Maria Brigida Ferraro, Francesca Martella
}
\examples{
data(butterfly)
}
\keyword{data}
\keyword{multivariate}
