\name{datasetsICR-package}
\alias{datasetsICR-package}
\docType{package}
\title{
The R Package datasetsICR}
\description{
This package contains some datasets analyzed in the book "An Introduction To Clustering With R".
}
\details{
For a complete list, use \code{library(help = "datasetsICR")}.
}
\references{
Giordani, P., Ferraro, M.B., Martella, F.: An Introduction to Clustering with R. Springer, Singapore (2020) 
}
\author{
Paolo Giordani, Maria Brigida Ferraro, Francesca Martella
}
