\name{Eurostat}
\alias{Eurostat}
\docType{data}
\title{
Eurostat dataset}
\description{
Economic indicators observed on some European countries in 2018.
}
\usage{data(Eurostat)}
\format{
A data.frame with 29 rows on 4 variables.
}
\details{
The dataset refers to 4 economic indicators for 29 European countries in 2018.
The indicators are: \cr
\code{Inflation}: Annual  HICP inflation rate (in percentage); \cr
\code{Poverty}: People at risk of poverty or social exclusion; \cr
\code{Unemployment}: Total unemployment rate; \cr
\code{GDP}: GDP per capita in PPS.
}
\source{
https://ec.europa.eu/eurostat/data/database
}
\references{
Giordani, P., Ferraro, M.B., Martella, F.: An Introduction to Clustering with R. Springer, Singapore (2020)
}
\author{
Paolo Giordani, Maria Brigida Ferraro, Francesca Martella
}
\examples{
data(Eurostat)
}
\keyword{data}
\keyword{multivariate}
