\name{FIFA}
\alias{FIFA}
\docType{data}
\title{
FIFA dataset}
\description{
Football analytics from the FIFA 19 database.
}
\usage{data(FIFA)}
\format{
A data.frame with 18207 rows on 80 variables.
}
\details{
The dataset contains the detailed attributes for every player registered in the latest edition of FIFA 19 database. Note that some player names display incorrectly because non-ASCII characters have been removed.
}
\source{
https://www.kaggle.com/karangadiya/fifa19
}
\references{
Gadiya, K.: FIFA 19 complete player dataset (2018) \cr
Giordani, P., Ferraro, M.B., Martella, F.: An Introduction to Clustering with R. Springer, Singapore (2020)
}
\author{
Paolo Giordani, Maria Brigida Ferraro, Francesca Martella
}
\examples{
data(FIFA)
}
\keyword{data}
\keyword{multivariate}