\name{USstate}
\alias{USstate}
\docType{data}
\title{
US state dataset}
\description{
FIPS codes for the US states
}
\usage{data(USstate)}
\format{
A data.frame with 50 rows on 3 variables.
}
\details{
The dataset refers to the FIPS codes for the US states. The variables are \code{fips} (FIPS State Numeric Code), \code{usps} (Official USPS Code) and \code{name} (Name).
}
\source{
https://www.census.gov/
}
\references{
Giordani, P., Ferraro, M.B., Martella, F.: An Introduction to Clustering with R. Springer, Singapore (2020)
}
\author{
Paolo Giordani, Maria Brigida Ferraro, Francesca Martella
}
\examples{
data(USstate)
}
\keyword{data}
\keyword{multivariate}